battle atk def Int32

+++ Core/GObject/Fighter.h
@@ -384,18 +384,18 @@ public:
 	inline float getExtraSoulP() { checkDirty(); return _attrExtraEquip.soulP; }
 	inline float getExtraAuraP() { checkDirty(); return _attrExtraEquip.auraP; }
 	inline float getExtraAuraMaxP() { checkDirty(); return _attrExtraEquip.auraMaxP; }
-    inline void addExtraAttack(UInt16 atk) { setDirty(true); _attrExtraEquip.attack += atk; }
-	inline UInt16 getExtraAttack() { checkDirty(); return _attrExtraEquip.attack; }
+    inline void addExtraAttack(Int32 atk) { setDirty(true); _attrExtraEquip.attack += atk; }
+	inline Int32 getExtraAttack() { checkDirty(); return _attrExtraEquip.attack; }
 	inline float getExtraAttackP() { checkDirty(); return _attrExtraEquip.attackP; }
-	inline void addExtraMagAttack(UInt16 atk) { setDirty(true); _attrExtraEquip.magatk += atk; }
-	inline UInt16 getExtraMagAttack() { checkDirty(); return _attrExtraEquip.magatk; }
+	inline void addExtraMagAttack(Int32 atk) { setDirty(true); _attrExtraEquip.magatk += atk; }
+	inline Int32 getExtraMagAttack() { checkDirty(); return _attrExtraEquip.magatk; }
 	inline float getExtraMagAttackP() { checkDirty(); return _attrExtraEquip.magatkP; }
-	inline UInt16 getExtraDefend() { checkDirty(); return _attrExtraEquip.defend; }
+	inline Int32 getExtraDefend() { checkDirty(); return _attrExtraEquip.defend; }
 	inline float getExtraDefendP() { checkDirty(); return _attrExtraEquip.defendP; }
-	inline UInt16 getExtraMagDefend() { checkDirty(); return _attrExtraEquip.magdef; }
+	inline Int32 getExtraMagDefend() { checkDirty(); return _attrExtraEquip.magdef; }
 	inline float getExtraMagDefendP() { checkDirty(); return _attrExtraEquip.magdefP; }
-	inline void addExtraHP(UInt32 hp) { setDirty(true); _attrExtraEquip.hp += hp; }
-	inline UInt16 getExtraHP() { checkDirty(); return _attrExtraEquip.hp; }
+	inline void addExtraHP(Int32 hp) { setDirty(true); _attrExtraEquip.hp += hp; }
+	inline Int32 getExtraHP() { checkDirty(); return _attrExtraEquip.hp; }
 	inline float getExtraHPP() { checkDirty(); return _attrExtraEquip.hpP; }
 	inline float getExtraAction() { checkDirty(); return _attrExtraEquip.action; }
 	inline float getExtraActionP() { checkDirty(); return _attrExtraEquip.actionP; }
@@ -463,10 +463,10 @@ public:
 	inline Int16 getBaseAuraMax() { return auraMax; }
 	inline Int16 getBaseTough() { return tough; }
 
-	inline Int16 getBaseAttack() { return attack; }
-	inline Int16 getBaseMagAttack() { return magatk; }
-	inline Int16 getBaseDefend() { return defend; }
-	inline Int16 getBaseMagDefend() { return magdef; }
+	inline Int32 getBaseAttack() { return attack; }
+	inline Int32 getBaseMagAttack() { return magatk; }
+	inline Int32 getBaseDefend() { return defend; }
+	inline Int32 getBaseMagDefend() { return magdef; }
 	inline float getBaseHitrate() { return hitrate; }
 
 	inline Int16 getBaseSoul() { return baseSoul; }
